+++ backend/lessons/scorm/engine/core/__init__.py
@@ -1,6 +1,8 @@
 from .abc import AbstractCore
 from .core import CoreSCORM
+from .base import BaseSCORMCore
 
 __all__ = ('AbstractCore',
            'CoreSCORM',
+           'BaseSCORMCore',
            )
