+++ backend/lessons/scorm/engine/core/abc.py
@@ -2,6 +2,9 @@ from abc import ABC, abstractmethod
 
 
 class AbstractCore(ABC):
+    """
+    Астрактрый класс для назначение протокола
+    """
 
     @abstractmethod
     def save(self): ...
